requirements.txt
PyOpenGL
PyOpenGL_accelerate
glfw
Pillow
numpy 